requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
brotli-python=1.0.9=py311hd77b12b_7
bzip2=1.0.8=h2bbff1b_5
ca-certificates=2024.3.11=haa95532_0
certifi=2024.2.2=py311haa95532_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cuda-cccl=12.4.127=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.127=0
cuda-opencl-dev=12.4.127=0
cuda-profiler-api=12.4.127=0
cuda-runtime=12.1.0=0
filelock=3.13.1=py311haa95532_0
freetype=2.12.1=ha860e81_0
gmpy2=2.1.2=py311h7f96b67_0
idna=3.4=py311haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
jinja2=3.1.3=py311haa95532_0
jpeg=9e=h2bbff1b_1
lerc=3.0=hd77b12b_0
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.3.5.147=0
libcurand-dev=10.3.5.147=0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_0
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libwebp-base=1.3.2=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.3=py311h2bbff1b_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py311h2bbff1b_1
mkl_fft=1.3.8=py311h2bbff1b_0
mkl_random=1.2.4=py311h59b6b97_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py311haa95532_0
networkx=3.1=py311haa95532_0
numpy=1.26.4=py311hdab7c0b_0
numpy-base=1.26.4=py311hd01c5d8_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.0.13=h2bbff1b_0
pillow=10.2.0=py311h2bbff1b_0
pip=23.3.1=py311haa95532_0
pysocks=1.7.1=py311haa95532_0
python=3.11.9=he1021f5_0
pytorch=2.2.2=py3.11_cuda12.1_cudnn8_0
pytorch-cuda=12.1=hde6ce7c_5
pytorch-mutex=1.0=cuda
pyyaml=6.0.1=py311h2bbff1b_0
requests=2.31.0=py311haa95532_1
setuptools=68.2.2=py311haa95532_0
sqlite=3.41.2=h2bbff1b_0
sympy=1.12=py311haa95532_0
tbb=2021.8.0=h59b6b97_0
tk=8.6.12=h2bbff1b_0
torchvision=0.17.2=pypi_0
typing_extensions=4.9.0=py311haa95532_1
tzdata=2024a=h04d1e81_0
urllib3=2.1.0=py311haa95532_1
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.41.2=py311haa95532_0
win_inet_pton=1.1.0=py311haa95532_0
xz=5.4.6=h8cc25b3_0
yaml=0.2.5=he774522_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
